+++ changelog.md
@@ -3,12 +3,16 @@
 ### New Bonuses
 - **Race (and racial subtype) Target**
   - Activate bonuses when the targeted tokens have one of the chosen races and/or subtypes.
+- **Spell Subschool Target**
+  - Activate bonuses when casting a spell from the given subschool(s).
 
 ### Misc
 - Dice Transform now also adds a new `@base` roll variable when using it--this is includes the whole "base dice" E.g. if you want to repeat `2d6` instead of having to do `(@quantity)d(@faces)` you can now do something like `@base + @quantity` to get `2d6 + 2`. See the in game documentation for more examples.
+- Replaced a few checkbox dialogs with the system's traits selector.
 
 ### Bugfixes
 - Furious Focus will now only trigger for melee attacks (i.e. no more triggering with Deadly Aim) and only with two-handed attacks.
+- Replaced a few missing untranslated strings
 
 ## Roll Bonuses 2.18.5